feat: split sub component and fix issue on form
