+++ packages/apps/project/src/components/molecules/SectionHeader/SectionHeader.tsx
@@ -1,0 +1,54 @@
+/*
+ * Axelor Business Solutions
+ *
+ * Copyright (C) 2024 Axelor (<http://axelor.com>).
+ *
+ * This program is free software: you can redistribute it and/or  modify
+ * it under the terms of the GNU Affero General Public License, version 3,
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import React from 'react';
+import {StyleSheet, View} from 'react-native';
+import {Icon, Text} from '@axelor/aos-mobile-ui';
+
+const SectionHeader = ({title, expanded, onPress, onEdit}) => (
+  <View style={styles.sectionContainer}>
+    <Text writingType="title">{title}</Text>
+    <View style={styles.iconContainer}>
+      <Icon
+        touchable={true}
+        onPress={onPress}
+        name={expanded ? 'chevron-up' : 'chevron-down'}
+        style={styles.icon}
+      />
+      <Icon touchable={true} onPress={onEdit} name="pencil-fill" />
+    </View>
+  </View>
+);
+
+const styles = StyleSheet.create({
+  sectionContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginHorizontal: 24,
+    marginVertical: 7,
+  },
+  iconContainer: {
+    flexDirection: 'row',
+  },
+  icon: {
+    marginRight: 10,
+  },
+});
+
+export default SectionHeader;
