+++ packages/apps/project/src/components/molecules/index.ts
@@ -25,5 +25,6 @@ export {default as DropdownTaskDescription} from './DropdownTaskDescription/Drop
 export {default as PartnerActionCard} from './PartnerActionCard/PartnerActionCard';
 export {default as ProjectHeader} from './ProjectHeader/ProjectHeader';
 export {default as ProjectSiteTags} from './ProjectSiteTags/ProjectSiteTags';
+export {default as SectionHeader} from './SectionHeader/SectionHeader';
 export {default as TaskActionCard} from './TaskActionCard/TaskActionCard';
 export {default as TaskDetailsHeader} from './TaskDetailsHeader/TaskDetailsHeader';
