+++ packages/apps/project/src/components/pages/CustomFieldFormView/CustomFieldFormView.tsx
@@ -18,66 +18,10 @@
 
 import React, {useMemo, useState, useCallback} from 'react';
 import {StyleSheet, View} from 'react-native';
-import {
-  CustomFieldForm,
-  useSelector,
-  useTranslator,
-} from '@axelor/aos-mobile-core';
-import {
-  HeaderContainer,
-  HorizontalRule,
-  Icon,
-  Text,
-  ScrollView,
-} from '@axelor/aos-mobile-ui';
+import {useSelector, useTranslator} from '@axelor/aos-mobile-core';
+import {HeaderContainer, ScrollView} from '@axelor/aos-mobile-ui';
 import {CustomFieldPopup, TaskDetailsHeader} from '../../molecules';
-
-const SectionHeader = ({title, expanded, onPress, onEdit}) => (
-  <View style={styles.sectionContainer}>
-    <Text writingType="title">{title}</Text>
-    <View style={styles.iconContainer}>
-      <Icon
-        touchable={true}
-        onPress={onPress}
-        name={expanded ? 'chevron-up' : 'chevron-down'}
-        style={styles.icon}
-      />
-      <Icon touchable={true} onPress={onEdit} name="pencil-fill" />
-    </View>
-  </View>
-);
-
-const CustomSection = ({
-  title,
-  expanded,
-  toggleSection,
-  onEdit,
-  fieldType,
-  modelId,
-  refreshKey,
-}) => (
-  <>
-    <SectionHeader
-      title={title}
-      expanded={expanded}
-      onPress={() => toggleSection(fieldType)}
-      onEdit={() => onEdit(fieldType, title)}
-    />
-    {expanded && (
-      <>
-        <CustomFieldForm
-          model="com.axelor.apps.project.db.ProjectTask"
-          fieldType={fieldType}
-          modelId={modelId}
-          style={styles.form}
-          readonly
-          key={refreshKey}
-        />
-        <HorizontalRule style={styles.horizontalRule} />
-      </>
-    )}
-  </>
-);
+import {CustomSection} from '../../organisms';
 
 const CustomFieldFormView = ({projecTaskId, config}) => {
   const I18n = useTranslator();
@@ -175,27 +119,6 @@ const styles = StyleSheet.create({
   },
   scrollView: {
     height: null,
-  },
-  form: {
-    flex: 1,
-  },
-  sectionContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginHorizontal: 24,
-    marginVertical: 7,
-  },
-  iconContainer: {
-    flexDirection: 'row',
-  },
-  icon: {
-    marginRight: 10,
-  },
-  horizontalRule: {
-    marginVertical: 12,
-    width: '90%',
-    alignSelf: 'center',
   },
 });
